[examples] Use lib built for example09 as pre-built lib for example08. Contribute to JB#44001
The example08 shows how to use a pre-built library in a project. In
order to allow the examples to be built cross-patform, we actually want
to use a library that's been built at the same time as the example.
Since this library comes from example09, this change allows us to use
the library built for example09 as the 'pre-built' library in example08.