[examples] Improve translation text and code style. Contribute to JB#44001
Following review, these changes are to generally improve the content and
style.

The section on translations is a bit terse for a first-time developer,
so links have been added to the Qt Linguist manual to give more context.

Some style and formatting issues with the code examples have been ironed
out.